backend/cfmp/requirements.txt
APScheduler==3.11.0
channels==4.2.2
cryptography==41.0.5
Django==5.2
django_apscheduler==0.7.0
django_minio_backend==3.8.0
django_minio_storage==0.5.8
djangorestframework==3.16.0
drf_extensions==0.8.0
minio==7.2.15
Pillow==11.2.1
PyJWT==2.10.1
PyJWT==2.10.1
urllib3==2.4.0
